Api documentation and fix errors

+++ backend/src/controllers/BookingController.js
@@ -29,9 +29,9 @@ export const getBooking = async( req, res ) => {
 
 export const createBooking = async( req, res) => {
     try {
-        const { from_date, to_date, price, transport, pet_id, branch_id } = req.body;
+        const { from_date, to_date, price, transport, comments, pet_id, branch_id } = req.body;
 
-        if( !from_date || !to_date || !price || !transport || !pet_id || !branch_id ){
+        if( !from_date || !to_date || !price || !pet_id || !branch_id ){
             return res.status( 400 ).json({ message: "El cuerpo de la solicitud está incompleto. Debes proporcionar todos los parámetros requeridos" }); 
         }
 
@@ -47,8 +47,7 @@ export const createBooking = async( req, res) => {
             return res.status(404).json({ message: 'Sucursal no encontrada' });
         }
 
-        const createdBooking = Booking.create({ from_date, to_date, price, transport, pet_id: pet.id, branch_id: branch.id })
-        
+        const createdBooking = await Booking.create({ from_date, to_date, price, transport, comments, pet_id: pet.id, branch_id: branch.id })
 
         return res.status( 201 ).json({ result: createdBooking });
     } catch (error) {
@@ -57,7 +56,7 @@ export const createBooking = async( req, res) => {
 }
 
 
-export const updateBooking = async() => {
+export const updateBooking = async(req, res) => {
     try {
         const { id } = req.params;
         
@@ -80,7 +79,7 @@ export const updateBooking = async() => {
     }
 }
 
-export const deleteBooking = async() => {
+export const deleteBooking = async(req, res) => {
     try {
         const { id } = req.params;
 
@@ -88,6 +87,11 @@ export const deleteBooking = async() => {
             return res.status( 400 ).json({ message: "El id es obligatorio" }); 
         }
 
+        const bookingToDelete = await Booking.findByPk( id )
+
+        if(!bookingToDelete){
+            return res.status( 404 ).json({ message: "La reserva no existe" }); 
+        }
         await Booking.destroy({
             where: {
                 id
@@ -98,4 +102,53 @@ export const deleteBooking = async() => {
     } catch ( error ) {
         return res.status( 500 ).json({ message: error.message });
     }
+}
+
+export const calculatePrice = async(req, res) => {
+    try {
+        const {booking_id , pet_id, days} = req.body
+        let price = 0
+
+        if(!booking_id || !pet_id || !days){
+            return res.status( 400 ).json({ message: "El cuerpo de la solicitud está incompleto. Debes proporcionar todos los parámetros requeridos" }); 
+        }
+        
+        const booking = await Booking.findByPk( booking_id,{
+            include: [
+                {
+                  model: Rate,
+                  as: 'rates',
+                }
+            ]}
+        )
+
+        if(!booking) {
+            return res.status( 404 ).json({ message: "La sucursal no existe" }); 
+        }
+
+        const pet = await Pet.findByPk( pet_id )
+
+        if(!pet) {
+            return res.status( 404 ).json({ message: "La mascota no existe" }); 
+        }
+
+        let petWeightInRange = false;
+        booking.rates.map(rate => {
+            if(pet.weight >= rate.weightFrom && pet.weight < rate.weightTo){
+                petWeightInRange = true;
+                price = rate.price * days
+            }
+        })
+
+        console.log(price)
+        if(!petWeightInRange){
+            return res.status( 404 ).json({ message: "No se encontro una tarifa para ese peso" }); 
+        }
+
+        return res.status( 200 ).json({ result: { price: price } })
+
+    } catch (error) {
+        return res.status( 500 ).json({ message: error.message });
+    }
+    
 }
